Added utilities/helper tests
