+++ src/components/Contact/Contact.test.tsx
@@ -2,29 +2,8 @@ import { act, fireEvent, render, screen } from '@testing-library/react';
 import axios from 'axios';
 import React from 'react';
 
-import { mockIntersectionObserver } from '../../utilities/mockIntersectionObserver';
+import { mockIntersectionObserver } from '../../mocks/mockIntersectionObserver';
 import { Contact } from './Contact';
-import { validateEmail } from './Contact.helper';
-
-const validEmails = [
-  '',
-  'foo@bar.com',
-  'foo.bar@baz.co.uk',
-  'jane@doe.test.gov',
-  'foo+bar@email.xyz',
-  'john-smith@baz.com',
-  'foo@bar-qux.io',
-];
-
-const invalidEmails = [
-  'foobar',
-  '#$%^&*(*&)',
-  '@test.com',
-  'apple.banana.com',
-  '.jane@doe.co.uk',
-  'foo..bar@baz.net',
-  '@',
-];
 
 describe('Contact component', () => {
   beforeEach(mockIntersectionObserver);
@@ -32,16 +11,6 @@ describe('Contact component', () => {
   it('renders', async () => {
     const { container } = render(<Contact />);
     expect(container).toMatchSnapshot();
-  });
-
-  it('validateEmail indentifies valid emails successfully', async () => {
-    for (const email of validEmails) {
-      expect(validateEmail(email)).toBe(true);
-    }
-
-    for (const email of invalidEmails) {
-      expect(validateEmail(email)).toBe(false);
-    }
   });
 
   it('contact from gives errors when required fields are empty', async () => {
